refactor(brutus): extract display functions to display.go
